requirements.txt
# Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Environment
python-dotenv==1.0.0

# API Requests
requests==2.31.0
openai==1.30.0

# Image Processing (install with: pip install --only-binary :all: pillow)
# Pillow - install manually if needed

# Video Processing
ffmpeg-python==0.2.0
opencv-python==4.9.0.80

# Database
SQLAlchemy==2.0.23

# Additional utilities
Werkzeug==3.0.1
click==8.1.7
